requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bokeh==2.4.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
compress-pickle==2.1.0
cycler==0.10.0
debugpy==1.5.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleads==3.8.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.20.1
hydralit==1.0.10
hydralit-components==1.0.5
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
jsonschema==4.0.1
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mediapipe==0.8.6
mistune==0.8.4
moviepy==1.0.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
playsound==1.3.0
plotly==5.3.1
praat-parselmouth==0.4.0
proglog==0.1.9
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.7.0
pygame==2.0.1
Pygments==2.10.0
PyJWT==2.2.0
pynput==1.7.3
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytube==11.0.1
pytz==2021.3
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
soupsieve==2.2.1
stopit==1.1.1
streamlit==1.0.0
suds-jurko==0.6
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
win10toast==0.9
wincertstore==0.2
wrapt==1.12.1
